+++ cr-20181201/src/Models/ListArtifactLifecycleRuleResponseBody/rules.php
@@ -22,7 +22,17 @@ class rules extends Model
     /**
      * @var bool
      */
+    public $dryRun;
+
+    /**
+     * @var bool
+     */
     public $enableDeleteTag;
+
+    /**
+     * @var bool
+     */
+    public $enableDeleteUntaggedManifest;
 
     /**
      * @var string
@@ -81,7 +91,9 @@ class rules extends Model
     protected $_name = [
         'auto' => 'Auto',
         'createTime' => 'CreateTime',
+        'dryRun' => 'DryRun',
         'enableDeleteTag' => 'EnableDeleteTag',
+        'enableDeleteUntaggedManifest' => 'EnableDeleteUntaggedManifest',
         'instanceId' => 'InstanceId',
         'modifiedTime' => 'ModifiedTime',
         'namespaceName' => 'NamespaceName',
@@ -114,8 +126,16 @@ class rules extends Model
             $res['CreateTime'] = $this->createTime;
         }
 
+        if (null !== $this->dryRun) {
+            $res['DryRun'] = $this->dryRun;
+        }
+
         if (null !== $this->enableDeleteTag) {
             $res['EnableDeleteTag'] = $this->enableDeleteTag;
+        }
+
+        if (null !== $this->enableDeleteUntaggedManifest) {
+            $res['EnableDeleteUntaggedManifest'] = $this->enableDeleteUntaggedManifest;
         }
 
         if (null !== $this->instanceId) {
@@ -188,8 +208,16 @@ class rules extends Model
             $model->createTime = $map['CreateTime'];
         }
 
+        if (isset($map['DryRun'])) {
+            $model->dryRun = $map['DryRun'];
+        }
+
         if (isset($map['EnableDeleteTag'])) {
             $model->enableDeleteTag = $map['EnableDeleteTag'];
+        }
+
+        if (isset($map['EnableDeleteUntaggedManifest'])) {
+            $model->enableDeleteUntaggedManifest = $map['EnableDeleteUntaggedManifest'];
         }
 
         if (isset($map['InstanceId'])) {