+++ cr-20181201/src/Models/UpdateArtifactLifecycleRuleRequest.php
@@ -16,7 +16,17 @@ class UpdateArtifactLifecycleRuleRequest extends Model
     /**
      * @var bool
      */
+    public $dryRun;
+
+    /**
+     * @var bool
+     */
     public $enableDeleteTag;
+
+    /**
+     * @var bool
+     */
+    public $enableDeleteUntaggedManifest;
 
     /**
      * @var string
@@ -59,7 +69,9 @@ class UpdateArtifactLifecycleRuleRequest extends Model
     public $tagRegexp;
     protected $_name = [
         'auto' => 'Auto',
+        'dryRun' => 'DryRun',
         'enableDeleteTag' => 'EnableDeleteTag',
+        'enableDeleteUntaggedManifest' => 'EnableDeleteUntaggedManifest',
         'instanceId' => 'InstanceId',
         'namespaceName' => 'NamespaceName',
         'repoName' => 'RepoName',
@@ -82,8 +94,16 @@ class UpdateArtifactLifecycleRuleRequest extends Model
             $res['Auto'] = $this->auto;
         }
 
+        if (null !== $this->dryRun) {
+            $res['DryRun'] = $this->dryRun;
+        }
+
         if (null !== $this->enableDeleteTag) {
             $res['EnableDeleteTag'] = $this->enableDeleteTag;
+        }
+
+        if (null !== $this->enableDeleteUntaggedManifest) {
+            $res['EnableDeleteUntaggedManifest'] = $this->enableDeleteUntaggedManifest;
         }
 
         if (null !== $this->instanceId) {
@@ -133,8 +153,16 @@ class UpdateArtifactLifecycleRuleRequest extends Model
             $model->auto = $map['Auto'];
         }
 
+        if (isset($map['DryRun'])) {
+            $model->dryRun = $map['DryRun'];
+        }
+
         if (isset($map['EnableDeleteTag'])) {
             $model->enableDeleteTag = $map['EnableDeleteTag'];
+        }
+
+        if (isset($map['EnableDeleteUntaggedManifest'])) {
+            $model->enableDeleteUntaggedManifest = $map['EnableDeleteUntaggedManifest'];
         }
 
         if (isset($map['InstanceId'])) {